Move the majority of shared types out of shared dir
