+++ fronts-client/src/shared/components/article/ArticleGraph.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { AreaChart, Area, XAxis, YAxis } from 'recharts';
-import { PageViewStory } from 'shared/types/PageViewData';
+import { PageViewStory } from 'types/PageViewData';
 import { theme } from '../../../constants/theme';
 import { State } from 'types/State';
 import { connect } from 'react-redux';
