+++ fronts-client/src/shared/types/Action.ts
@@ -1,8 +1,8 @@
-import { ExternalArticle } from './ExternalArticle';
+import { ExternalArticle } from '../../types/ExternalArticle';
 import { Collection, Card, Group, CardMeta } from './Collection';
 import { Actions } from 'lib/createAsyncResourceBundle';
 import { copyCardImageMeta } from 'shared/actions/Cards';
-import { PageViewStory } from './PageViewData';
+import { PageViewStory } from '../../types/PageViewData';
 import { PersistMeta } from 'util/storeMiddleware';
 
 interface CardsReceived {
