+++ fronts-client/src/types/ExternalArticle.ts
