+++ fronts-client/src/types/PageViewData.ts
