+++ fronts-client/src/util/CAPIUtils.ts
@@ -1,5 +1,5 @@
 import { Element, Tag, CapiArticle } from 'types/Capi';
-import { ExternalArticle } from '../shared/types/ExternalArticle';
+import { ExternalArticle } from '../types/ExternalArticle';
 import { CardMeta } from '../shared/types/Collection';
 import { notLiveLabels, liveBlogTones } from 'constants/fronts';
 import startCase from 'lodash/startCase';
