+++ fronts-client/src/util/form.ts
@@ -5,7 +5,7 @@ import clamp from 'lodash/clamp';
 import pickBy from 'lodash/pickBy';
 import { isDirty } from 'redux-form';
 import { CardMeta } from 'shared/types/Collection';
-import { DerivedArticle } from 'shared/types/Article';
+import { DerivedArticle } from 'types/Article';
 import { CapiArticle } from 'types/Capi';
 import { State } from 'types/State';
 import { selectCard, selectSharedState } from 'shared/selectors/shared';
